Check foliations
It turns out that since CGAL stores vertices and cells but not
edges and faces, it's easier to just grab a cell handle and check
the vertex->info() directly.

If we're going to do that, might as well classify the cells
as (4,0), (3,1), (2,2) and the like while we're at it.

This is just a start.