Fixed longstanding bug in copy ctor and move assignment
Non-static data members being initialized in the order in which they
are declared combined with using values led to a subtle bug. The move
assignment operator (and move ctor in the previous commit) relied on
this behavior, but then tried to do operations on the moved-from data
member to derive values from another , which was naturally null.

The copy ctor didn’t really have a bug, but it was called within the
context of the Metropolis call operator, which did. Namely, it too
attempted to return from an x-value, instead of the data structure it
contained.

But LLVM’s Function_ref.h worked like a charm compared to
std::function, which choked on the input.