Canonical class (#32)
* Disable non-relevant failing tests for now

* Require same version of CMake as CGAL 4.9; cdt-opt stub

* Start worker pattern

* Mock out simulation interface

* (Ab)using lambdas for nicer workflow

cdt-opt compiles but doesn’t run. Track down segfault there.

* Discovered yet another bug

The SimplicialManifold copy constructor throws EXC_BAD_ACCESS when
invoked from cdt-opt

* Eliminated 3 calls to SimplicialManifold copy ctor; added CTests

Borrowed LLVM’s function_ref, an efficienct type-erasing, non-owning
reference to a callable as a replacement for (inefficient)
std::function. Now just need to fix segfault in cdt-opt; added CTest
that shows this. Removed more testing artifacts in .gitignore.

* Fix cppcheck and an out-of-range test

* Debugging

Somewhere in the call chain for cdt-opt, the SimplicialManifold is
nulled out

* Refactor SimplicialManifold

Break out into its own file for clarity and further work

* Fixed SimplicialManifold move ctor bug!

* Fixed longstanding bug in copy ctor and move assignment

Non-static data members being initialized in the order in which they
are declared combined with using values led to a subtle bug. The move
assignment operator (and move ctor in the previous commit) relied on
this behavior, but then tried to do operations on the moved-from data
member to derive values from another , which was naturally null.

The copy ctor didn’t really have a bug, but it was called within the
context of the Metropolis call operator, which did. Namely, it too
attempted to return from an x-value, instead of the data structure it
contained.

But LLVM’s Function_ref.h worked like a charm compared to
std::function, which choked on the input.

* Documentation prior to merge