Use option types and rvalue references where needed
facet_timevalues is a std::set, and a std::set.insert() methods wants
an r-value. Also, use constant iterators for unchanging values.
std::set timevalues should be an option type because it is not
initialized when constructed, but only contains a value as a result of
the VolumePerTimeslice function. So, calls to max_timevalue() which
read from std::set timevalues should return 0 if timevalues hasn’t been
initialized.