Fix Codecov uploads
Codecov uploads didn't work when TESTS=true, so change back to TEST=false. Also change in Travis-CI, and verify that the CI build will pass even when the Codecov job fails, which it will with the hidden tests added back. Set the range for Codecov reports to the level of coverage we're currently at (and figure out how to include or ignore subrepos with/without tests of their own). Also update the README with the Codecov badge, and set notifications to work with failed CI jobs.