Fixing interface to Manifold
Ergodic moves should not have to know details about FoliatedTriangulations, Geometry, or other data structures in the Manifold. So, adding functions in Manifold that perfectly forward to the appropriate functions lower down in the data structure hierarchy. This sets the stage to re-write FoliatedTriangulations, which need to know a lot of data currently in Geometry in order to check that they are correctly constructed. This in turn breaks Ergodic Moves unless they have been properly insulated from these changes, which they will if they only have to refer to functions in Manifold.