Foliated Triangulation refactor, part 1
A Foliated Triangulation has-a Delaunay Triangulation, hence private inheritance. Moving containers and algorithms out of Geometry and into FoliatedTriangulation, so that checks on data structures can be done much sooner. Also, Geometry will just contain scalar values, so it will be easy to move/copy when we speculatively execute moves that have calculations that only depend on numbers.