Better detection of invalid cells
check_timeslices does a better job of detecting invalidly foliated cells, and outputs a container of those vertices to be deleted. If there is an equal number of high and low, badly foliated vertices, prefer to delete lower timevalues so that the desired number of timeslices is more likely to be obtained. 289 tests passed.