check_vertices now detects timevalues out of range
Check_vertices access the CGAL::Compact_container which has the vertices directly. The result is not a vertex handle, so we get the timevalue of the infinite vertex and use that to determine if a vertex in the compact container is the infinite vertex. We ignore it in our tests if so, because the compact container returns all vertices, including the infinite one. We test the remaining vertices to see if they're within bounds of the maximum timevalue and the minimum timevalue. 288 tests passed.