Uncomment move command tests
After a lengthy refactoring of the base data structures, return
to handling moves sequentially, atomically, and with rollback
functionality using the Command and Memento patterns. This version
fails, next check-in should pass. Also update documentation and
remove unnecessary tests. Expect 6 tests to fail of 288 from tests
on the compact container expecting to get a vertex.