Explore failure modes of Move_command
Calling update on the result of a move command violates the precondition on collect_cells (which is called when the manifold is updated), and ms-gsl calls std::terminate. But the move appears to be actually happening, as it does without use of Move_command. Uncomment line 171 of Move_command_test to see the fireworks. 280/284 passing tests.