Pre-fix an upcoming bug with CGAL+TBB
CGAL_LINKED_WITH_TBB is not being set, and manually setting it in Settings.hpp produces linker errors, perhaps because CGAL code does not also have it turned on. So, in an upcoming version of CGAL the Triangulation_data_structure_3 does not default to Sequential if CGAL::Parallel_tag is set but CGAL_LINKED_WITH_TBB is not. Thanks to @jefftrull for the heads up! 326/336 tests passing.