Fix Foliated_triangulation swap
Use CGAL's Triangulation_3 swap function for swapping the Delaunay_triangulation_3 data member of a Foliated_triangulation<3>.  This affects the copy/move assignment operators and the move constructor of a Foliated_triangulation and therefore, the swap function of a Manifold which is defined in terms of the Foliated_triangulation swap. CGAL assumes that the swapped_from Triangulation_3 is immediately discarded.