Refactor leaky abstractions
The data structures and operations were tangled and confused. A Geometry object encapsulates numeric data for rapid retrieval by the algorithms which calculate the action and probability of moves. A FoliatedTriangulation encapsulates a Delaunay triangulation, but with useful methods and class members to ensure it is correctly initialized. A Manifold handles both Geometry and FoliatedTriangulations in-flight, while simulations are running, ensuring that both are (lazily) consistent as necessary. So move methods and data around to better fit this model. Also throwaway old data structures.