(feature, ci) Does vertex radius match timevalue?
Add a function to check the squared distance of a vertex from the origin and compare it to the squared timevalue. These should match for proper vertices. This required adding new data members to remember the initial radius and spacing of the foliations in the Foliated Triangulation.

Because we are comparing squared distances using doubles, add a TOLERANCE interval for the comparison.

Also change Travis-CI to require TSAN to pass.

425/431 tests passing.