(fix) Formulas and functions for radius and expected values thereof
Now that the formula for the radius is fixed, the expected timevalue can be calculated nearly exactly and returned as an integer, making comparisons with the vertex timevalue (which is stored as an int) precise.

Also, in Tetrahedron tests we use points that have the correct spacing for the foliation, and use the new functions to check correctness.

This will lead to checking the status of all vertices and simplices as part of initialization.

437/440 tests passing.