(ci) AppVeyor's CGAL upgrade, part 2
AppVeyor caches things, so when the old formula uses a library (mpir) that conflicts with a newer one (gmp), you have to uninstall it. But if you uninstall it after you've "installed" everything (which is typically a no-op, because they're cached) then when you go to upgrade, the libraries you want are no longer there to be upgraded. So, you have to uninstall mpir before you install, then upgrade.