(fix) Finding and fixing incorrect vertices works now
Now that we can find and fix incorrect vertices, the next step is to fix incorrect cells that are mis-classified as a result of incorrect vertices. We may also need to look at the foliation logic, to make sure that it's catching all the errors. We should definitely incorporate the fix_cells and fix_vertices as part of building the respective containers.

So far, we've made the decision to build additional containers into the Foliated Triangulation rather than just use the lower level half-edge data structures in the Triangulation. The rationale for this is the half-edge is fast and efficient and concerns itself mainly with connectivity. The foliation of the triangulation is not something it knows about, so that should be handled by the Foliated Triangulation class.