(build) Enable vcpkg manifest mode
It is easier and better to keep track of project dependencies inside vcpkg.json and allow CMake to automatically build them. However, this does cause CMake to invoke vcpkg on every build, which takes a minute or two even with local caching. So leave vcpkg in classic mode for CI.

Also set $VCPKG_ROOT which in turn sets $CMAKE_TOOLCHAIN_FILE automatically, which is a general improvement.