(build,ci) Use vcpkg Manifest mode
Thanks to Robert Schumacher and the #vcpkg channel in the Cpplang Slack, we can now use the manifest to build vcpkg ports one time only and have all builds re-use them. This is helpful since vcpkg builds 100+ libraries and it takes several minutes to do so, even with local caching, if the default manifest mode of building in each build directory is used.

For now, we'll continue to use Classic mode in all the CI builds until we can verify the workflow.

Also standardize on setting VCPKG_ROOT in all CI.