(ci) Fix manifest mode CI errors
Travis-CI and GitHub Actions build fine, so this adds caching and similar workflow.

AppVeyor can't yet use manifest mode because it requires installation of yasm-tool:x86-windows, and there does not appear to be a way to do that with manifests, yet. So AppVeyor stays in classic mode.

Still need to fix Docker images to use new workflow.