(refactor) FoliatedTriangulations
Fix suggestions by cppcheck.

In general, functions in the FoliatedTriangulation namespace should be generic (i.e. templated), and work on triangulations of any dimensionality. Functions specific to an instantiation of a FoliatedTriangulation should go in the template specialization.

There is still some cleanup to do: adding tests for 4D functions, and generalizing others in the FoliatedTriangulations namespace to any dimensionality.

Also, having to specify 2 template parameters is ugly, perhaps add a few user-defined deduction guides.