(ci) Fix SonarCloud workflow, part 3 [ci skip]
Fix setting the executable path for build-wrapper-linux-x86-64 to GitHub Actions environment variable. I define the location of executables at the beginning of the job, because they are cached. Therefore, we want to skip any downloading and unzipping on cache hits. This means we can't set the environment variable during that step, in case we have a cache hit. Thus set it for the entire job; a cache hit will find it, and a cache miss will create it.