(feature) spdlog for classic mode debugging
When the debugger ignores breakpoints, sometimes the old "print out a statement here" is best. So we picked a robust tool to do it. Dipping toes in, starting with basics (and ensuring compatibility with various platforms and compilers).

Also switch back to CodeCov bash uploader, because codecov/codecov-action@v2 isn't working.