(feature) create_logger
Log messages according to the general hierarchy:

* Trace: Used to trace internals of a function
* Debug: Diagnostic output on a function
* Info: General information
* Warn: Errors that are handled
* Err: Errors which cause a function to fail
* Critical: Errors which cause the program to fail

Log Info and below to the console, Debug and below to logs/debug-log.txt, and everything to logs/trace-log.txt.

Change the default build to RelWithDebInfo, because Debug will now log a *lot* of information if all tests are run. Gigabyte sized debug-log.txt are not uncommon.

Thanks to https://stackoverflow.com/questions/2031163/when-to-use-the-different-log-levels