(fix) Debugging!
Fixed a long-standing issue with interactive debugging on CLion.

https://intellij-support.jetbrains.com/hc/en-us/requests/3057187

Interactive debugging breakpoints are disabled if LTO is enabled, so turn that off by default. But Release and RelWithDebInfo builds can turn it back on.

Catch otherwise uncaught exceptions propagating in tests, even if clang-tidy and other static analyzers don't see it.