(refactor) fix_timeslices -> fix_timevalues part 1
A cell is acausal if 1) it spans more than 1 timeslice 2) it is completely on a single timeslice. All the information to determine if the triangulation has acausal simplices can be mostly obtained from the cells with functions already present. The function check_timevalues uses expected_cell_type to return a container of invalid cells. We use std::optional in order to allow the semantics that a false value means there are no acausal cells, otherwise a container of acausal cells is returned.

However, in order to delete the invalid vertices we'll need a key-value pair of the vertices in a given acausal cell, hence the function vertices_from_cell.

Then a future fix_timevalues function can use  check_timevalues to get the incorrect cells and vertices_from_cell to apply logic to determine which vertices to delete to fix the triangulation.