(refactor) fix_timeslices -> fix_timevalues part 2
Put basic outline in place. Needs unit tests to start refining. Big surprise, this is really a transform-reduce, so it's been restructured accordingly.

We input a container of invalid cells (from check_cells), transform them into a container of vertices to delete, and then reduce that to just the unique vertices (it's possible the same invalid vertex is in multiple cells).