(feature) Foliated triangulations can now fix themselves
Completed refactoring the free functions for detecting and fixing errors in simplices, cells, and the foliation of the Delaunay triangulation. Added a member function to FoliatedTriangulation that calls these.

Also update vcpkg commit ID in remaining GitHub Actions, along with a new cache key on SonarCloud, which forces a download of the latest version of sonar-scanner and it's build-wrapper.