(fix) Bug in (2,6) move
The function find_26_move (renamed to find_adjacent_31_cell) looks for (2,6) move sites by finding the (3,1) cell adjacent to the (1,3) cell that it starts from. It was originally looking at all neighbors via the cell index, which means in a boundary cell or when there are only 2 it could erroneously find and accept a cell that was adjacent to the infinite vertex, which nevertheless somehow is misclassified as a (3,1) cell. This was found doing the minimal triangulation setup of a (3,1) cell stacked on a (1,3) cell.

This bug was fixed by looking at the expected cell type of the neighbor, rather than trusting the cell->info() data which could be misclassified.

Continue to hunt down and remove for loops.

Fixed the scan.sh script, which relies on hardcoded paths to the location of llvm installed by Homebrew, because locate scripts generate too many results.

Also move CodeCov back to Travis-CI because it worked there but does not work on GitHub Actions. Disable running the corresponding GitHub Action on push and pull requests.

Need to debug the (6,2) move next. This type of bug is probably why some of the cells attempted look erroneous, and why the majority of test failures occur with the (2,6) and (6,2) moves. Also need to resolve the misclassification of cells adjacent to the infinite vertex.