(test, ci) Added test of (6,2) move on minimal triangulation
All the tests pass, except the correctness of the manifold, which is disturbing. There's something odd going on with cells being classified incorrectly. The naive attempt (commented out) to simply fix the vertices and cells doesn't yet work.

Also attempt to fix ci issues on Travis and GitHub Actions. For Travis-CI, we need to let the tests run for CodeCov longer. For GitHub Actions on Windows, we updated vcpkgGitCommitId to the most recent passing commit, but probably I'll file a bug report when this doesn't resolve the tbb installation issue.

Update the cache key for SonarCloud so it gets the latest build-wrapper.