(test) Fixed Ergodic move tests for (2,6) and (6,2) moves
An off-by-one error when check that the foliation is correct. Now confident that moves work as intended.

This allows all vertices to be checked for correctness when looking for (6,2) move sites.

This is probably slow, and should be replaced with a function that only checks the vertices at the proposed move site.