(feature) Concepts and constraints
Due to the deep class hierarchy, some operations are constrained by the types of containers they can execute on. C++20 Concepts document this.

The std::shuffle algorithm as used in the move functions requires that the container they operate on have swappable elements. This is satisfied by std::vector but not, in all cases by std::span (for certain compilers and operating systems, unfortunately).