(fix) Really fix bug
Turns out I was recording initial moves twice; once during Metropolis::initialize and then when running the remaining moves with Metropolis::operator(). The previous fix removed updating the Metropolis counters during initialization, but that causes those tests to fail (which means we're not aware of the moves made during Metrpolis::initialize when it returns a MoveCommand). The solution is to reset the counters of the MoveCommand before returning it in Metropolis::initialize.

I could have also just used 2 different MoveTrackers, but I'm going with the functional flow of the Metropolis::initialize returning a MoveCommand that's been properly initialized.

Also fiddled around clang-format to get brace alignments more to taste, which of course changed most files.