(refactor) Make write_file easier to use
Add static class variable so that the topology of the manifold can be queried, which allows the generation of a filename be done with just the members of the Manifold class. This allows the write_file to take the Manifold as an argument and from that generate the filename and save appropriately. 

This also fixes a bug in CDT where the initial radius and foliation spacing were always set to 1.0, instead of reading the values from the manifold.