(ci) More GitHub Actions and AppVeyor tuning with CMakePresets.json
Clang-15 seems the closest to getting to work with lcov, so switch to that.

It also now seems to work on Ubuntu 22.04.

Made a cmake preset for AppVeyor.

Turns out the cmake --preset=build also builds packages in vcpkg, so the Install packages step was installing them twice! That should speed up builds considerably on GitHub Actions.