(ci) Also remove the tests for cdt-viewer [ci skip]
On GitHub Actions using macOS, we want to disable building cdt-viewer. Which worked, but forgot to also disable the tests.

Unfortunately this only saves link/compile time for cdt-viewer, but not the 3 hours it takes to build qt on the GitHub Actions runner with macOS.

vcpkg.json only understands platform, not environment variables.