(test) Write a bistellar-flip test harness
We want to draw a minimal triangulation before and after a bistellar flip. The reason this is not in /tests is because it uses CGAL::draw() to display the results, which requires Qt5. Qt5 is extremely heavy and not needed for the other unit tests, so I don't want to require/link it to the test binary.