(fix) Enforce FoliatedTriangulation class invariant
The FoliatedTriangulation class invariant is that the timevalues and cell types stored in the vertices and cells of the Delaunay triangulation must correct.

The initialization order of the data members was re-arranged so that the triangulation, foliation initial radius, and foliation spacing were initialized first. This allows the vertices to be checked and corrected, and they are initialized next. Once the vertices have the correct timevalues, the cells can be initialized with the correct cell types. All other containers of edges and faces (sub-simplices) derive from the vertices and cells.

Also update to vcpkgGitCommitId: ae59b93af0c1dce4126e963fca5b330662495fa1