Add back default cases and remove Expects and Ensures
Removing the default case from switches triggers a warning on Sonarcloud.

Ensures and Expects just end up calling std::terminate, but without any indication as to where the error is. So replace these with assert which will at least print the error line in Debug and be removed with NDEBUG.