Use a span of Vertex_handles for read-only
The commit introduces the use of a span of `Vertex_handle` objects for read-only access in the code. It modifies the `Foliated_triangulation.hpp` and `Manifold.hpp` files to add a new member function `get_vertices_span()` that returns a span of vertices. The changes also include updating the test cases in `Vertex_test.cpp` to use the new member function.